refactor for learning utils
