+++ adam/learner/__init__.py
@@ -1,15 +1,31 @@
 """
 Interfaces for language learning code.
 """
-
+from adam.learner.language_mode import LanguageMode
 from abc import ABC, abstractmethod
 from pathlib import Path
-from typing import Dict, Generic, Mapping, Optional, Callable, Tuple
+from typing import (
+    Dict,
+    Generic,
+    Mapping,
+    Optional,
+    Callable,
+    Tuple,
+    Iterable,
+    List,
+    AbstractSet,
+)
 from immutablecollections import ImmutableSet, immutableset
 from adam.learner.alignments import (
     LanguagePerceptionSemanticAlignment,
     PerceptionSemanticAlignment,
 )
+from adam.learner.surface_templates import (
+    STANDARD_SLOT_VARIABLES,
+    SurfaceTemplate,
+    SurfaceTemplateBoundToSemanticNodes,
+)
+from enum import Enum, auto
 import itertools
 from adam.learner.alignments import LanguageConceptAlignment
 from adam.semantics import SemanticNode
@@ -228,42 +244,37 @@ class ComposableLearner(ABC):
         """
 
 
-@attrs(frozen=True, slots=True)
-class SemanticNodeWithSpan:
-    """This is a tuple class currently used by our verb and relation learners"""
-
-    node: SemanticNode = attrib(validator=instance_of(SemanticNode))
-    span: Span = attrib(validator=instance_of(Span))
-
-
-def in_left_to_right_order(semantic_nodes: Tuple[SemanticNodeWithSpan, ...]) -> bool:
-    """This function checks that a sequence is in left to right order"""
-    previous_node = semantic_nodes[0]
-    for i in range(1, len(semantic_nodes)):
-        if not previous_node.span.precedes(semantic_nodes[i].span):
-            return False
-        previous_node = semantic_nodes[i]
-    return True
-
-
-def aligned_object_nodes(
-    num_arguments: int,
-    num_arguments_to_alignments_sets: Dict[
-        int, ImmutableSet[Tuple[SemanticNodeWithSpan, ...]]
-    ],
+def covers_entire_utterance(
+    bound_surface_template: SurfaceTemplateBoundToSemanticNodes,
     language_concept_alignment: LanguageConceptAlignment,
-) -> ImmutableSet[Tuple[SemanticNodeWithSpan, ...]]:
-    if num_arguments not in num_arguments_to_alignments_sets.keys():
-        # we haven't seen a request for this number of arguments before so we need to generate all the valid options
-        semantic_nodes_with_spans = immutableset(
-            SemanticNodeWithSpan(node=node, span=span)
-            for (node, span) in language_concept_alignment.node_to_language_span.items()
-        )
-        num_arguments_to_alignments_sets[num_arguments] = immutableset(
-            ordered_semantic_nodes
-            for ordered_semantic_nodes in itertools.product(
-                semantic_nodes_with_spans, repeat=num_arguments
+    *,
+    ignore_determiners: bool = False,
+) -> bool:
+    num_covered_tokens = 0
+    for element in bound_surface_template.surface_template.elements:
+        if isinstance(element, str):
+            num_covered_tokens += 1
+        else:
+            num_covered_tokens += len(
+                language_concept_alignment.node_to_language_span[
+                    bound_surface_template.slot_to_semantic_node[element]
+                ]
             )
-            if in_left_to_right_order(ordered_semantic_nodes)
+    # We may need to ignore counting english determiners in our comparison
+    # to the template as the way we treat english determiners is currently
+    # a hack. See: https://github.com/isi-vista/adam/issues/498
+    sized_tokens = (
+        len(language_concept_alignment.language.as_token_sequence())
+        if not ignore_determiners
+        else len(
+            [
+                token
+                for token in language_concept_alignment.language.as_token_sequence()
+                if token not in ["a", "the"]
+            ]
         )
-    return num_arguments_to_alignments_sets[num_arguments]
+    )
+
+    # This assumes the slots and the non-slot elements are non-overlapping,
+    # which is true for how we construct them.
+    return num_covered_tokens == sized_tokens
